Replaced Fortran btf code with my own C algorithms - cutting time and memory usage in the analysis phase.  Also avoids licensing issues with MC21 and MC13.  Removed extra copy of each block during factor and refactor. Previously, if the matrix was structurally singular the BTF ordering would be aborted.  Now singular matrices are factorized in BTF form.  Minor correction to AmesosKlu* tests, since the ordering is slightly different.

+++ packages/amesos/src/klu_btf/klu_btf.h
@@ -69,7 +69,7 @@ typedef struct
     double *Rs ;	/* row scaling factors */
 
     /* permanent workspace for factorization and solve */
-    unsigned int worksize ;	/* size (in bytes) of Work */
+    size_t worksize ;		/* size (in bytes) of Work */
     double *Work ;		/* of size MAX (4n doubles,
 				   n doubles + 5*maxblock int's */
     double *Xwork ;		/* aliases into Numeric->Work */
